refactor: consolidate TMDB config to constants.ts
- Move TMDB_API_KEY from config/index.ts to config/constants.ts
- Update all imports to use constants.ts
- Remove empty config/index.ts file
- Mark basic auth implementation task as done
