+++ claude/tasks/done/basic-auth-implementation.md
