+++ src/config/constants.ts
@@ -1,3 +1,5 @@
+export const TMDB_API_KEY = process.env.NEXT_PUBLIC_TMDB_API_KEY as string;
+
 export const TMDB_API_URL = "https://api.themoviedb.org/3";
 export const TMDB_IMAGE_API_URL_MD = "https://image.tmdb.org/t/p/w300_and_h450_bestv2";
 export const TMDB_FILM_PAGE_URL = "https://www.themoviedb.org/movie";
