--- src/config/index.ts
@@ -1,1 +1,0 @@
-export const TMDB_API_KEY = process.env.NEXT_PUBLIC_TMDB_API_KEY as string;
